enrichment-metrics-function/requirements.txt
azure-functions
pyodbc
azure-identity